front end/requirements.txt
# Frontend requirements
streamlit==1.28.0
pandas==2.1.3
numpy==1.24.3
openai==0.28.0
requests==2.31.0
python-dotenv==1.0.0
pytesseract==0.3.10
Pillow==10.1.0
pdf2image==1.16.3
PyPDF2==3.0.1
plotly==5.17.0
streamlit-chat==0.0.2
pymupdf==1.23.8
streamlit-option-menu==0.3.6